decouple ice<->review relation, fix tests

+++ src/reviews/reviews.service.ts
@@ -5,48 +5,41 @@ import { Model } from 'mongoose';
 import { Review, ReviewDocument } from './entities/review.entity';
 import { OurExceptionType } from 'src/common/errors/OurExceptionType';
 import { OurHttpException } from 'src/common/errors/OurHttpException';
-import { UsersService } from 'src/users/users.service';
-import { IceCreamsService } from 'src/ice-creams/ice-creams.service';
 import { CheckReviewDto } from './dto/check-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
+import { RankingStatus } from './types/RankingStatus';
 
 @Injectable()
 export class ReviewsService {
   constructor(
     @InjectModel(Review.name) private reviewModel: Model<ReviewDocument>,
-    private readonly iceCreamService: IceCreamsService,
   ) {}
 
-  async updateReviewAndUpdateRating(dto: UpdateReviewDto, reviewId: string) {
+  async updateReview(dto: UpdateReviewDto, reviewId: string) {
     const review = await this.getReviewById(reviewId);
-    console.log(`revie out:${review}`);
 
     review.content = dto.content;
     review.lastUpdate = new Date();
+
     if (!dto.rating) {
       return await review.save();
     }
-    console.log(
-      await this.iceCreamService.updateIceCreamRatingAfterUpdatedReview(review),
-    );
+
     review.rating = dto.rating;
     return await review.save();
   }
 
-  async createReviewAndUpdateRating(dto: CreateReviewDto) {
+  async createReview(dto: CreateReviewDto) {
     const review = await this.reviewModel.findOne({
       userId: dto.userId,
       iceCreamId: dto.iceCreamId,
     });
+
     if (review) {
       throw new OurHttpException(OurExceptionType.REVIEW_ALREADY_EXISTS);
     }
 
     const createdReview = new this.reviewModel(dto);
-
-    await this.iceCreamService.updateIceCreamRatingAfterNewReview(
-      createdReview,
-    );
 
     return await createdReview.save();
   }
@@ -79,10 +72,29 @@ export class ReviewsService {
     return !review ? false : true;
   }
 
-  async removeReviewAndUpdateRating(id: string) {
-    const review = await this.reviewModel.findById(id);
+  async getIceCreamRankingStatus(iceCreamId: string): Promise<RankingStatus> {
+    const reviews = await this.getIceCreamAllReviews(iceCreamId);
+    if (!reviews)
+      return {
+        averageRating: 0,
+        numberOfReviews: 0,
+      };
 
-    await this.iceCreamService.updateIceCreamRatingAfterDeletedReview(review);
+    let totalRating = reviews.reduce(
+      (sum: number, review: Review) => sum + review.rating,
+      0,
+    );
+    return {
+      averageRating: totalRating / reviews.length,
+      numberOfReviews: reviews.length,
+    };
+  }
+
+  async removeReview(id: string) {
+    const review = await this.reviewModel.findById(id);
+    if (!review) {
+      throw new OurHttpException(OurExceptionType.REVIEW_DOES_NOT_EXIST);
+    }
     await this.reviewModel.deleteOne({ _id: id });
   }
 }
